updating didc version
